Separate success into multiple files
[MAILPOET-2142]
